fixes #4 implement assign attendees to a counselor
Implement better selection of counselors assigned to attendees.
Add a method to unassign a counselor and assign another counselor to an attendee.
Add a default message to a few error response methods.
Change respondUnprocessableEntity to response with the error response.
Remove unnecessary methods from AttendeesController.
Add a controller method to assign a new counselor to an attendee.
Add a database dependency for dates.
Add tests around new functionality.
Add more assertions to AttendeesTest.